+++ src/VCAuthn/IdentityServer/Endpoints/AuthorizationEndpoint/AuthorizationEndpointResult.cs
