+++ src/VCAuthn/IdentityServer/Endpoints/AuthorizationEndpoint/AuthorizeEndpoint.cs
@@ -9,7 +9,6 @@ using IdentityServer4.Validation;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using VCAuthn.ACAPy;
 using VCAuthn.IdentityServer.SessionStorage;
 using VCAuthn.PresentationConfiguration;
 using VCAuthn.UrlShortener;
