+++ src/VCAuthn/PresentationConfiguration/PresentationRecord.cs
@@ -3,25 +3,6 @@ using Newtonsoft.Json;
 
 namespace VCAuthn.PresentationConfiguration
 {
-    public class PresentationRequest
-    {
-        [JsonProperty("@id")]
-        public string Id { get; set; }
-
-        [JsonProperty("@type")]
-        public string Type => "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/credential-presentation/0.1/presentation-request";
-
-        [JsonProperty("request")]
-        public PresentationConfiguration Request { get; set; }
-        
-        [JsonProperty("comment")]
-        public string Comment { get; set; }
-        
-        [JsonProperty("thread_id")]
-        public string ThreadId { get; set; }
-    }
-    
-    
     public class PresentationRecord
     {
         [JsonProperty("id")]
